fix compile and remove unnecessary MetaStoreFS interface
